requirements.txt
fastapi
uvicorn
django
djangorestframework
django-filter
markdown
ipython
pdbpp